Update patch set 1
Patch Set 1:

The prebuilt app is/was simply temporary until I got a chance to figure out how to add it to the build system.
The prebuilt was a quick fix solution to get the app in the same nightly that was getting USB host and such.

The source is on github already

Patch-set: 1