Update patch set 1
Patch Set 1:

I am sorry for the late reply. I was under an impression that this patch was merged already.

Both the OMX.TI.Video.Decoder and OMX.TI.720P.Decoder utilize the DSP to process the video frames. All that I have done is to change the order of the Decoders, so that the generic video decoder is selected first and if it fails, then the 720p decoder is selected. Previously, 720p decoder was being selected for all the videos irrespective of the codec capability that the video requires, which would totally defeat the purpose of building and including a generic TI video decoder, since it was never used. 

We don't always require the 720p decoder (ITTIAM) unless the resolution of the target video is high. The generic TI Video decoder does the job well and it's memory footprint is a lot less when compared to 720P decoder. 

Also, a maximum of 4 instances of generic TI video decoder could run at a time, whereas the only one instance of 720p decoder could run at a time on OMAP3 due to the memory requirement constraints. Hence, having the generic TI video decoder selected first would improve the performance in Gallery wherein you have a lot of non 720P videos, since the thumbnail generation would be much faster considering 4 instances of generic video decoders are created to decode the video and generate thumbnails rather than having the 720p decoder selected always and slowing down the operation due to a single instance limitation.

https://github.com/CyanogenMod/android_hardware_ti_omap3/commit/c355703424efb47ee24233f57d71d258a87fb28d

Patch-set: 1
Reviewer: Dheeraj CVR <3545@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0